migrate content of work_item_relation.py
